refactor: rename to slug
