+++ src/app/product/[slug]/layout.tsx
